--- busy_beaver/apps/github_integration/summary/__init__.py
@@ -1,1 +1,0 @@
-from . import task  # noqa
